app/src/main/livekit_requirements.txt
livekit-agents>=0.8.0
livekit-plugins-openai>=0.8.0
livekit-plugins-silero>=0.8.0
python-dotenv>=1.0.0 